requirements.txt
asgiref==3.8.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
defusedxml==0.7.1
Django==5.2.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
numpy==2.3.0
oauthlib==3.3.0
orjson==3.10.18
packaging==24.2
propcache==0.3.2
pycparser==2.22
PyJWT==2.9.0
PyPDF2==3.0.1
python-decouple==3.8
python3-openid==3.2.0
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
tenacity==9.1.2
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
zstandard==0.23.0
